No float ever touches money or stored quantities in app/
Quarterly tax totals move out of the view into TaxPayment.quarter_total
(exact BigDecimal); customer revenue, expense totals, and gallons sums
convert to BigDecimal. The ratios that legitimately stay float — MPG and
payment rate — now say so in place.

+++ app/controllers/expenses_controller.rb
@@ -13,8 +13,8 @@ class ExpensesController < ApplicationController
     @expenses = @expenses.where("expense_date >= ?", @start_date) if @start_date
     @expenses = @expenses.where("expense_date <= ?", @end_date) if @end_date
 
-    @expense_by_category = @expenses.group(:category).sum(:amount).sort_by { |_k, v| -v.to_f }
-    @expense_total = @expenses.sum(:amount).to_f
+    @expense_by_category = @expenses.group(:category).sum(:amount).sort_by { |_k, v| -v.to_d }
+    @expense_total = @expenses.sum(:amount).to_d
     @expenses = @expenses.order(expense_date: :desc)
 
     respond_to do |format|
